Introduce bbclass srcuri-force-nobranch
Some upstream projects rename their default branch from master to main,
which causes builds to break. This bbclass provides a workaround using
the nobranch option in case updating the affected layer is not an option.

The nobranch option disabled the validation of the SRCREV against the
branch given for the specific URI within SRC_URI, or the fallback value.

Please refer to the official Yocto documentation for more details on this
option.

Signed-off-by: Alexander Herfurtner <alexander.h@variscite.com>